requirements.txt
langchain>=0.0.217
duckduckgo-search
streamlit-feedback
assemblyai
openai
langchain_nvidia_ai_endpoints
langchain_community
faiss-cpu
python-dotenv
streamlit
pypdf
haystack
assemblyai-haystack
sentence-transformers
huggingface_hub
pytube
requests
PyPDF2==3.0.1
reportlab
fpdf